backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
Flask-Limiter==3.8.0
python-socketio==5.9.0
python-engineio==4.7.1
python-dotenv==1.0.0
requests==2.31.0
supabase==2.9.1
python-dateutil==2.8.2
PyPDF2==3.0.1
pdfminer.six
Pillow
playwright==1.40.0
PyJWT==2.9.0
bcrypt==4.2.1
gunicorn==21.2.0
eventlet==0.37.0
